third_party/WebKit/LayoutTests/typedcssom/computedstyle/pseudo-elements-expected.txt

This is a testharness.js-based test.
FAIL get on Computed StyleMap of pseudo element returns correct styles assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
PASS getAll on Computed StyleMap of pseudo element returns list of correct styles 
Harness: the test ran to completion.
